fix: clippy lints, let-chain refactors, and minor correctness improvements
Addresses clippy warnings across multiple crates and modernizes
conditional patterns using Rust let-chains for improved readability.

Key changes:
- ftui-extras: remove unnecessary parentheses in clamp_f64() calls
  throughout text_effects.rs and StyledMultiLine (9 instances)
- ftui-render: use let-chains in grapheme_pool refcount increment,
  remove unnecessary usize cast from DiffRun::len() calls in diff.rs
  and presenter.rs cost model
- ftui-text: replace `% 2 == 0` with `.is_multiple_of(2)` in bidi.rs
  cursor positioning; remove unnecessary &str-to-&str borrows in wrap.rs
- ftui-widgets: refactor nested if-let chains to flat let-chains in
  file_picker.rs (symlink confinement), list.rs (filtered index cache),
  and table.rs (display index cache); add CachedTableDisplayIndices
  type alias; fix status_line.rs right-segment underflow with .max(area.x)

+++ crates/ftui-harness/tests/isomorphism_proofs.rs
@@ -859,7 +859,7 @@ fn proof_runs_cover_all_changes() {
     let changes = diff.changes();
 
     // Count total cells covered by runs
-    let run_cells: usize = runs.iter().map(|r| r.len() as usize).sum();
+    let run_cells: usize = runs.iter().map(|r| r.len()).sum();
 
     assert_eq!(
         run_cells,
